LPS-146474 Splitted constants folder in DXPActivationKeysTable container
