+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/containers/DXPActivationKeysTable/utils/constants/activationKeysLicenseFilterTypes.js
@@ -1,0 +1,18 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * The contents of this file are subject to the terms of the Liferay Enterprise
+ * Subscription License ("License"). You may not use this file except in
+ * compliance with the License. You can obtain a copy of the License by
+ * contacting Liferay, Inc. See the License for the specific language governing
+ * permissions and limitations under the License, including but not limited to
+ * distribution rights of the Software.
+ */
+
+export const ACTIVATION_KEYS_LICENSE_FILTER_TYPES = {
+	active: ({expirationDate, startDate}) =>
+		new Date(startDate) < new Date() &&
+		new Date(expirationDate) > new Date(),
+	expired: ({expirationDate}) => new Date(expirationDate) < new Date(),
+	notActivated: ({startDate}) => new Date(startDate) > new Date(),
+};
