+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/containers/DXPActivationKeysTable/utils/constants/activationStatus.js
@@ -1,0 +1,33 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * The contents of this file are subject to the terms of the Liferay Enterprise
+ * Subscription License ("License"). You may not use this file except in
+ * compliance with the License. You can obtain a copy of the License by
+ * contacting Liferay, Inc. See the License for the specific language governing
+ * permissions and limitations under the License, including but not limited to
+ * distribution rights of the Software.
+ */
+
+export const ACTIVATION_STATUS = {
+	active: {
+		color: 'success',
+		id: 'active',
+		title: 'Active',
+	},
+	all: {
+		color: 'none',
+		id: 'all',
+		title: 'All',
+	},
+	expired: {
+		color: 'danger',
+		id: 'expired',
+		title: 'Expired',
+	},
+	notActivated: {
+		color: 'info',
+		id: 'notActivated',
+		title: 'Not Activated',
+	},
+};
