+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/containers/DXPActivationKeysTable/utils/constants/columns-definitions/EnvironmentTypeColumn.js
@@ -1,0 +1,36 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * The contents of this file are subject to the terms of the Liferay Enterprise
+ * Subscription License ("License"). You may not use this file except in
+ * compliance with the License. You can obtain a copy of the License by
+ * contacting Liferay, Inc. See the License for the specific language governing
+ * permissions and limitations under the License, including but not limited to
+ * distribution rights of the Software.
+ */
+
+import {getPascalCase} from '../../../../../utils/getPascalCase';
+
+const EnvironmentTypeColumn = ({activationKey}) => {
+	const productName = getPascalCase(activationKey.licenseEntryType)
+		.split('-')
+		.join(' ');
+
+	const productDescription = activationKey.complimentary
+		? 'Complimentary'
+		: 'Subscription';
+
+	return (
+		<div>
+			<p className="font-weight-bold m-0 text-neutral-10">
+				{productName.replace('Production', 'Prod')}
+			</p>
+
+			<p className="font-weight-normal m-0 text-neutral-7 text-paragraph-sm">
+				{productDescription}
+			</p>
+		</div>
+	);
+};
+
+export {EnvironmentTypeColumn};
