+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/containers/DXPActivationKeysTable/utils/constants/columns.js
@@ -1,0 +1,70 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * The contents of this file are subject to the terms of the Liferay Enterprise
+ * Subscription License ("License"). You may not use this file except in
+ * compliance with the License. You can obtain a copy of the License by
+ * contacting Liferay, Inc. See the License for the specific language governing
+ * permissions and limitations under the License, including but not limited to
+ * distribution rights of the Software.
+ */
+
+import ClayIcon from '@clayui/icon';
+
+export const COLUMNS = [
+	{
+		accessor: 'envName',
+		bodyClass: 'border-0',
+		expanded: true,
+		header: {
+			description: 'Description',
+			name: 'Environment Name',
+			styles: 'bg-transparent',
+		},
+	},
+	{
+		accessor: 'keyType',
+		bodyClass: 'border-0',
+		header: {
+			description: 'Host Name / Cluster Size',
+			name: 'Key Type',
+			noWrap: true,
+			styles: 'bg-transparent',
+		},
+	},
+	{
+		accessor: 'envType',
+		bodyClass: 'border-0',
+		header: {
+			name: 'Environment Type',
+			styles: 'bg-transparent text-neutral-10 font-weight-bold',
+		},
+	},
+	{
+		accessor: 'expirationDate',
+		bodyClass: 'border-0',
+		header: {
+			name: 'Exp. Date',
+			styles: 'bg-transparent text-neutral-10 font-weight-bold',
+		},
+		noWrap: true,
+	},
+	{
+		accessor: 'status',
+		align: 'center',
+		bodyClass: 'border-0',
+		header: {
+			name: 'Status',
+			styles: 'bg-transparent text-neutral-10 font-weight-bold',
+		},
+	},
+	{
+		accessor: 'download',
+		align: 'center',
+		bodyClass: 'border-0',
+		header: {
+			name: <ClayIcon symbol="download" />,
+			styles: 'bg-transparent text-neutral-10 font-weight-bold',
+		},
+	},
+];
